[Kashif]Refactored:To take in List of Census Data instead of Iterating through the File
