--- src/main/java/commonscsv/CSVBuilderFactory.java
@@ -1,8 +1,0 @@
-package commonscsv;
-
-public class CSVBuilderFactory {
-
-	public static <E> ICSVBuilder<E> createCSVBuilder() {
-		return new CommonsCSVBuilder<E>();
-	}	
-}
