--- src/main/java/commonscsv/CSVException.java
@@ -1,15 +1,0 @@
-package commonscsv;
-
-public class CSVException extends Exception {
-
-	public enum ExceptionType {
-		INCORRECT_STATE;
-	}
-
-	ExceptionType type;
-
-	public CSVException(String message, ExceptionType type) {
-		super(message);
-		this.type = type;
-	}
-}
