--- src/main/java/commonscsv/ICSVBuilder.java
@@ -1,10 +1,0 @@
-package commonscsv;
-
-import java.io.Reader;
-import java.util.Iterator;
-
-public interface ICSVBuilder<E> {
-
-	public Iterator<E> getCSVIterator(Reader reader, Class<E> csvClass) throws CSVException ;
-
-}
